Rename preconditions class
