+++ core/src/test/java/io/spine/change/BooleanMismatchShould.java
