+++ core/src/test/java/io/spine/change/ChangesShould.java
