+++ core/src/test/java/io/spine/change/IntMismatchShould.java
