+++ core/src/test/java/io/spine/change/StringMismatchShould.java
